refactor: remove old Lookup pages
